% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_agg}
\alias{data_agg}
\title{data_agg}
\format{
A tibble with 288 rows and 8 variables:
\describe{
  \item{keyword}{\code{character} keyword for which the data was downloaded}
  \item{date}{date as \code{integer} variable, can be transformed into date
  format with \code{lubridate::as_date}}
  \item{type}{\code{character} indicating the type of timeseries-column from
  \code{data_score} that is used for DOI computation, takes either
  "socre_obs", "score_sad", or "score_trd"}
  \item{gini}{\code{double} DOI computed as inverted Gini coefficient of
  search score from \code{data_score}}
  \item{hhi}{\code{double} DOI computed as inverted Herfindahl-Hirschman
  index of search score from \code{data_score}}
  \item{entropy}{\code{double} DOI computed as inverted Entropy measure for
  search score from \code{data_score}}
  \item{batch_c}{\code{integer} number of control batch}
  \item{batch_o}{\code{integer} number of object batch}
  \item{lst_geo}{\code{character} list of locations for which the search 
  score is used}
}
}
\usage{
data_agg
}
\description{
xxx.
}
\keyword{datasets}
