% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.r
\name{.export_data_single}
\alias{.export_data_single}
\alias{.export_data_double}
\title{Export data from database tables with single batch}
\usage{
.export_data_single(
  table,
  in_keyword = NULL,
  in_object = NULL,
  in_control = NULL,
  in_locations = NULL
)

.export_data_double(
  table,
  in_keyword = NULL,
  in_object = NULL,
  in_control = NULL,
  in_locations = NULL,
  in_type = NULL
)
}
\description{
Export data from database tables with single batch

Export data from database tables with double batches
}
\keyword{internal}
