% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_data.r
\name{.remove_batch_keywords}
\alias{.remove_batch_keywords}
\alias{.remove_batch_time}
\alias{.remove_data_control}
\alias{.remove_data_object}
\alias{.remove_data_mapping}
\alias{.remove_data_score}
\alias{.remove_data_doi}
\alias{.remove_data_global}
\title{Remove from batch_keywords}
\usage{
.remove_batch_keywords(type, batch)

.remove_batch_time(type, batch)

.remove_data_control(batch)

.remove_data_object(batch)

.remove_data_mapping(batch_c = NULL, batch_o = NULL)

.remove_data_score(batch_c = NULL, batch_o = NULL)

.remove_data_doi(batch_c = NULL, batch_o = NULL)

.remove_data_global(batch)
}
\description{
Remove from batch_keywords

Remove from batch_time

Remove from data_control

Remove from data_object

Remove from data_mapping

Remove from data_score

Remove from data_doi

Remove from data_global
}
\keyword{internal}
