% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_global.r
\name{download_global}
\alias{download_global}
\alias{download_global.numeric}
\alias{download_global.list}
\title{Download global score data for object batch}
\usage{
download_global(object)

\method{download_global}{numeric}(object)

\method{download_global}{list}(object)
}
\arguments{
\item{object}{Object batch for which the data is downloaded. Object
of class \code{numeric}.}
}
\value{
Message that data was downloaded successfully. Data is uploaded
to data_wrld.
}
\description{
Download global score data for object batch
}
\details{

}
\examples{
\dontrun{
data_wrld(object = 1)
data_wrld(object = as.list(1:5))
}

}
\seealso{

}
