% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mapping.r
\name{download_mapping}
\alias{download_mapping}
\alias{download_mapping.numeric}
\alias{download_mapping.list}
\title{Download data for mapping between control and object batch}
\usage{
download_mapping(control, object, locations = lst_wdi)

\method{download_mapping}{numeric}(control, object, locations = lst_wdi)

\method{download_mapping}{list}(control, object, locations = lst_wdi)
}
\arguments{
\item{control}{Control batch for which the data is downloaded. Object
of class \code{numeric}.}

\item{object}{Object batch for which the data is downloaded. Object
of class \code{numeric} or object of class \code{list} containing single
elements of class \code{numeric}.}

\item{locations}{List of countries or regions for which the data is downloaded.
Refers to lists generated in \code{start_db}.}
}
\value{
Message that data was downloaded successfully. Data is uploaded
to data_map.
}
\description{
Download data for mapping between control and object batch
}
\details{

}
\examples{
\dontrun{
data_map(control = 1, object = 1, locations = lst_wdi)
data_map(control = 1, object = as.list(1:5), locations = lst_wdi)
}

}
\seealso{

}
