% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_change.r
\name{export_voi_change}
\alias{export_voi_change}
\alias{export_doi_change}
\title{Export changes in data}
\usage{
export_voi_change(keyword = NULL, object = NULL, control = NULL, type = "obs")

export_doi_change(
  keyword = NULL,
  object = NULL,
  control = NULL,
  locations = NULL,
  type = NULL,
  measure = "gini"
)
}
\arguments{
\item{keyword}{Object keywords for which data should be exported. Object of
class \code{character}.}

\item{object}{Object batch number for which data should be exported.}

\item{control}{Control batch number for which data should be exported.}

\item{type}{Type of time series for which data should be exported. Element
of class \code{character}. Relevant only for \code{export_global} and
\code{export_doi}. Takes one of the following values: \emph{obs} - observed
search scores, \emph{sad} - seasonally adjusted search scores, \emph{trd} -
trend only search scores.}

\item{locations}{List of locations for which the search score is used.
For \code{export_control}, \code{export_object}, or \code{export_score}
refers to lists generated in \code{start_db}. For \code{export_doi}
object of class \code{character}.}

\item{measure}{Object of class \code{character} indicating the measure used
for DOI computation for which abnormal changes should be analyzed. Takes
either \emph{gini}, \emph{hhi}, or \emph{entropy}. Defaults to \emph{"gini"}.}
}
\value{
The functions export and filter the respective database tables and return
objects of class \code{"tbl_df", "tbl", "data.frame"}.
\itemize{
  \item \code{export_voi_change} exports data from table \emph{data_score}
  with columns keyword, date, control, object, voi, voi_change, quantile,
  filters for \code{location == "world"}.
  \item \code{export_doi_change} exports data from table \emph{data_doi} with
  columns keyword, date, type, control, object, locations, doi, doi_change,
  quantile.
}
}
\description{
The function allows to export changes in search scores, voi, and doi and
shows percentile of changes to identify abnormal changes. In combination with
various \emph{write} functions in R, the functions allow exports from the
database to local files.
}
\details{
The function computes changes in VOI or DOI between dates. To identify
abnormal changes, the function computes the percentile rank for each change
within the distribution of changes. Low percentile rank signify abnormally
high negative changes. High percentile ranks signify abnormally high positive
changes.
Exports can be filtered by \emph{keyword}, \emph{object}, \emph{control},
\emph{locations}, \emph{type}, or \emph{measure}. Not all filters are
applicable for all functions. When filter \emph{keyword} and \emph{object}
are used together, \emph{keyword} overrules \emph{object}. Currently the
functions do not include list inputs - users are advised to
\code{purrr::map_dfr} or \code{dplyr::filter} instead.
}
\examples{
\dontrun{
export_voi_change(
  keyword = "amazon",
  type = "obs"
)

export_doi_change(
  keyword = "amazon",
  locations = "countries",
  type = "obs",
  measure = "gini"
)

# interaction with purrr::map_dfr
purrr::map_dfr(
  c("coca cola", "microsoft"),
  export_voi_change,
  control = 1,
  type = "obs"
)

# interaction with dplyr::filter
export_doi_change(
  object = 1,
  control = 1,
  locations = "countries",
  type = "obs",
  measure = "gini"
) \%>\%
  dplyr::filter(lubridate::year(date) == 2019)
}

}
\seealso{
* \code{\link{export_doi}}
* \code{\link[purrr]{map}}
* \code{\link[dplyr]{filter}}
}
