% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_change.r
\name{export_doi_change}
\alias{export_doi_change}
\title{Export changes in data}
\usage{
export_doi_change(
  keyword = NULL,
  object = NULL,
  control = NULL,
  locations = NULL,
  type = NULL,
  measure = "gini"
)
}
\arguments{
\item{keyword}{Object keywords for which data should be exported. Object of
class \code{character}.}

\item{object}{Object batch number for which data should be exported.}

\item{control}{Control batch number for which data should be exported.}

\item{locations}{List of locations for which the search score is used.
For \code{export_control}, \code{export_object}, or \code{export_score}
refers to lists generated in \code{start_db}. For \code{export_doi}
object of class \code{character}.}

\item{type}{Type of time series for which data should be exported. Element
of class \code{character}. Relevant only for \code{export_global} and
\code{export_doi}. Takes one of the following values: \emph{obs} - observed
search scores, \emph{sad} - seasonally adjusted search scores, \emph{trd} -
trend only search scores.}

\item{measure}{Object of class \code{character} indicating the measure used
for DOI computation for which abnormal changes should be analyzed. Takes
either \emph{gini}, \emph{hhi}, or \emph{entropy}. Defaults to \emph{"gini"}.}
}
\value{

}
\description{
The function allows to export changes in search scores, voi, and doi and
shows percentile of changes to identify abnormal changes.
}
\details{

}
\examples{
\dontrun{
export_doi_change(keyword = "amazon", type = "obs", locations = "countries")
}

}
\seealso{
* \code{\link{export_doi}}
* \code{\link[purrr]{map}}
* \code{\link[dplyr]{filter}}
}
