% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtrends_base_db.r
\name{gtrends_base}
\alias{gtrends_base}
\title{Load doiGT database and tables}
\usage{
gtrends_base()
}
\value{
The function exports the following objects to .GlobalEnv:
\itemize{
  \item gtrends_db A DBIConnection object, as returned by
  \code{DBI::dbConnect()}, connecting to the SQLite database in the working
  directory
  \item data_geo A remote data source pointing to the table "data_geo" in
  the connected SQLite database
  \item batch_terms A remote data source pointing to the table "batch_terms"
  in the connected SQLite database
  \item batch_time A remote data source pointing to the table "batch_times"
  in the connected SQLite database
  \item data_agg A remote data source pointing to the table "data_agg" in
  the connected SQLite database
  \item data_con A remote data source pointing to the table "data_con" in
  the connected SQLite database
  \item data_map A remote data source pointing to the table "data_map" in
  the connected SQLite database
  \item data_obj A remote data source pointing to the table "data_obj" in
  the connected SQLite database
  \item data_score A remote data source pointing to the table "data_score" in
  the connected SQLite database
  \item data_wrld A remote data source pointing to the table "data_wrld" in
  the connected SQLite database
  \item lst_wdi A \code{character} vector containing ISO2 country codes of
  countries that add at leas 0.1% to global GDP
  \item lst_usa A \code{character} vector containing ISO2 regional codes of
  US states
  \item terms_con A \code{tibble} containing keywords of control batches
  \item time_con A \code{tibble} containing times of control batches
  \item terms_obj A \code{tibble} containing keywords of object batches
  \item time_obj A \code{tibble} containing times of control batches
  \item dict_obj A \code{tibble} containing synonymous keywords
}
}
\description{
Load doiGT database and tables
}
\examples{
\dontrun{
gtrends_base()
}

}
\seealso{

}
