% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_doi.r, R/compute_score.r,
%   R/database_functions.r, R/helper_functions.r
\name{.compute_gini}
\alias{.compute_gini}
\alias{.compute_hhi}
\alias{.compute_entropy}
\alias{.reset_date}
\alias{.adjust_ts}
\alias{.aggregate_synonym}
\alias{.enter_location}
\alias{.get_trend}
\alias{.test_batch}
\alias{.test_empty}
\title{Compute gini coefficient}
\usage{
.compute_gini(series)

.compute_hhi(series)

.compute_entropy(series)

.reset_date(data)

.adjust_ts(data)

.aggregate_synonym()

.enter_location(globaltrends_db)

.get_trend(location, term, time = "all")

.test_batch(batch)

.test_empty(
  table,
  batch_c = NULL,
  batch_o = NULL,
  location = NULL,
  locations = NULL
)
}
\description{
Compute gini coefficient

Compute herfindahl hirschman index

Compute entropy

Reset date

Adjust time series

Aggregate synonyms

Enter location data into database

Download search trends from Google

Test batch number

Test whether table has entries for given criteria
}
\keyword{internal}
