% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.r
\name{plot_bar}
\alias{plot_bar}
\alias{plot_bar.exp_score}
\alias{plot_score_bar}
\alias{plot_bar.abnorm_score}
\alias{plot_abnorm_score_bar}
\title{Create barplot for cross-sectional globaltrends data}
\usage{
plot_bar(data, ...)

\method{plot_bar}{exp_score}(data, type = "obs", ...)

plot_score_bar(data, type = "obs")

\method{plot_bar}{abnorm_score}(data, ...)

plot_abnorm_score_bar(data)
}
\arguments{
\item{data}{Data exported from \code{export_...} or \code{compute_abnorm} functions.}

\item{...}{Further arguments passed to or from other methods.}

\item{type}{Object of type \code{character} indicating the type of time
series-column from data_score, takes either \emph{obs}, \emph{sad}, or \emph{trd}. Defaults
to \emph{"obs"}.}
}
\value{
Barplot of cross-sectional data as \code{ggplot2} object.
}
\description{
The function creates barplots for cross-sectional search score data. It uses
the output of \code{export_score} to prepare a bar plot of search scores for
the top 10 countries. For output from \code{get_abnorm_hist} the plot shows
five locations with the highest and lowest abnormal changes each. When the
output includes more than one keyword, only the first keyword is used.
}
\details{
The \code{plot_xxx_bar} functions allow to call \code{plot_bar} on a data object that
lost its respective class (e.g., due to a join).
}
\examples{
\dontrun{
data <- export_score(keyword = "amazon")
plot_bar(data, type = "obs")

# for cases where data looses the respective class
data <- export_score(keyword = "amazon")
plot_score_bar(data)

data <- export_score(keyword = "amazon")
data <- get_abnorm_hist(data, train_win = 12, train_break = 0, type = "obs")
plot_bar(data)
}

}
