% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_change.r
\name{plot_voi_change}
\alias{plot_voi_change}
\alias{plot_doi_change}
\title{Plot changes in data}
\usage{
plot_voi_change(data_change, ci = 0.95)

plot_doi_change(data_change, type = "obs", ci = 0.95)
}
\arguments{
\item{data_change}{Data exported from \code{export_voi_change} or
\code{export_doi_change} function.}

\item{ci}{Confidence interval within which changes are assumed to be normal.
Object of class \code{double, 0 < ci < 1}. Defaults to \emph{0.95}.}

\item{type}{Type of time series for which data should be exported. Element
of class \code{character}. Relevant only for \code{export_global} and
\code{export_doi}. Takes one of the following values: \emph{obs} - observed
search scores, \emph{sad} - seasonally adjusted search scores, \emph{trd} -
trend only search scores.}
}
\value{
Line plot of change time series as \code{ggplot2} object. Changes with
percentile ranks outside the provided confidence interval are indicated by
red dots.
}
\description{
The function uses the output of \code{export_voi_change} and
\code{export_doi_change} to prepare a time series plot of changes in VOI or
DOI. When the output includes more than one keyword, only the first keyword
is used. Dots indicate abnormal changes in the underlying time series.
}
\examples{
\dontrun{
export_voi_change(
  keyword = "amazon",
  type = "obs"
)
plot_voi_change(data_change = data)

data <- export_doi_change(
  keyword = "amazon",
  locations = "countries",
  type = "obs",
  measure = "gini"
)
plot_doi_change(data_change = data)
}

}
\seealso{
* \code{\link{export_doi_change}}
* \code{\link{export_voi_change}}
* \code{\link[purrr]{map}}
* \code{\link[dplyr]{filter}}
}
