% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_doi_box.r
\name{plot_doi_box}
\alias{plot_doi_box}
\title{Boxplot of DOI time series}
\usage{
plot_doi_box(data_doi, type = "obs", measure = "gini", locations = "countries")
}
\arguments{
\item{data_doi}{Data exported from \code{export_doi} function.}

\item{type}{Object of class \code{character} indicating the type of time
series-column from data_score that is used for DOI computation, takes
either "obs", "sad", or "trd". Defaults to \code{"obs"}.}

\item{measure}{Object of class \code{character} indicating the measure
used for DOI computation, takes either "gini", "hhi", or "entropy".
Defaults to "gini".}

\item{locations}{Object of class \code{character} indicating for which
set of locations should be filtered. Defaults to \code{"countries"}.}
}
\value{
Boxplot of time series as \code{ggplot2} object.
}
\description{
The function uses the output of \code{export_doi} to prepare a box plot of
the distribution of degree of internationalization values. When the output
includes more than nine keywords, only the first nine keywords are used.
}
\examples{
\dontrun{
data <- export_doi(object = 1, locations = "countries")
plot_doi_box(data_doi = data, type = "obs", measure = "gini")
plot_doi_box(data_doi = data, type = "sad", measure = "hhi")
plot_doi_box(data_doi = data, type = "trd", measure = "entropy")
}

}
\seealso{
\code{\link{export_doi}}
}
