% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score.r
\name{plot_score}
\alias{plot_score}
\title{Bar plot of data_score distribution}
\usage{
plot_score(data_score, type = "obs")
}
\arguments{
\item{type}{Object of class \code{character} indicating the type of time
series-column from data_score that is used, takes either "obs", "sad", or
"trd". By default takes "obs".}
}
\value{
Bar plot of score distribution across locations as \code{ggplot2} object
}
\description{
Bar plot of data_score distribution
}
\details{

}
\section{Warning}{

\code{plot_score} is limited to 1 unique keyword to avoid an
overcrowded plot. When \code{data_doi} or \code{data_score} includes
more than 1 unique keyword, only the first keyword is used.
}

\examples{
\dontrun{
data <- export_score(keyword = "manchester united", locations = "countries")
plot_score(data_score = data, type = "obs")
plot_score(data_score = data, type = "sad")
data <- dplyr::filter(data, location \%in\% countries)
plot_score(data_score = data, type = "trd")
}

}
