% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend.r
\name{plot_trend}
\alias{plot_trend}
\title{Line plot of data_doi and data global time series}
\usage{
plot_trend(
  data_doi,
  data_score_global,
  type = "obs",
  measure = "gini",
  locations = "countries",
  smooth = TRUE
)
}
\arguments{
\item{data_doi}{Data exported from \code{export_doi} function.}

\item{data_score_global}{Data exported from \code{export_score_global}
function.}

\item{type}{Object of class \code{character} indicating the type of time
series-column from data_score that is used for DOI computation, takes
either "obs", "sad", or "trd". Defaults to \code{"obs"}.}

\item{measure}{Object of class \code{character} indicating the measure
used for DOI computation, takes either "gini", "hhi", or "entropy".
Defaults to "gini".}

\item{locations}{Object of class \code{character} indicating for which
set of locations should be filtered. Defaults to  \code{"countries"}.}

\item{smooth}{Object of class \code{logical} indicating whether the
\code{geom_smooth} function of \code{ggplot2} should be used. Defaults to
\code{TRUE}.}
}
\value{
Line plot of time series as \code{ggplot2} object
}
\description{
Line plot of data_doi and data global time series
}
\details{

}
\section{Warning}{

\code{plot_trend} is limited to 1 unique keyword to avoid an
overcrowded plot. When \code{data_doi} or \code{data_score} includes
more than 1 unique keyword, only the first keyword is used.
}

\examples{
\dontrun{
data1 <- export_doi(keyword = "manchester united", locations = "countries")
data2 <- export_score_global(keyword = "manchester united")
plot_trend(data_doi = data1, data_score_global = data2, type = "obs", measure = "gini", smooth = TRUE)
plot_trend(data_doi = data1, data_score_global = data2, type = "sad", measure = "hhi", smooth = FALSE)
plot_trend(data_doi = data1, data_score_global = data2, type = "trd", measure = "entropy", smooth = TRUE)
}

}
