% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.r
\name{plot_ts}
\alias{plot_ts}
\alias{plot_ts.exp_score}
\alias{plot_ts.exp_voi}
\alias{plot_ts.exp_doi}
\title{Plot time series of globaltrends data}
\usage{
plot_ts(data, ...)

\method{plot_ts}{exp_score}(data, smooth = TRUE, ...)

\method{plot_ts}{exp_voi}(data, smooth = TRUE, ...)

\method{plot_ts}{exp_doi}(
  data,
  measure = c("gini", "hhi", "entropy"),
  locations = "countries",
  smooth = TRUE,
  ...
)
}
\arguments{
\item{data}{Data exported from \code{export_...} or \code{compute_abnorm} functions.}

\item{...}{Further arguments passed to or from other methods.}

\item{smooth}{Object of type \code{logical} indicating whether the \code{geom_smooth}
function of \code{ggplot2} should be used. Defaults to \code{TRUE}.}

\item{measure}{Object of type \code{character} indicating the DOI measure,
takes either \emph{gini}, \emph{hhi}, or \emph{entropy}. Defaults to \emph{"gini"}.}

\item{locations}{Object of type \code{character} indicating for which
set of locations should be filtered. Defaults to \emph{"countries"}.}
}
\value{
Line plot of time series as \code{ggplot2} object. For plots for output from
\code{get_abnorm_hist} the provided confidence interval is indicated by red dots.
}
\description{
The function creates line plots for time series globaltrends data. It uses
the output of \code{export_...} to prepare line plots for up to nine keywords.
}
\details{
For output of \code{export_score}, only data for a single location is shown.
When date for more than one location is provided, the function selects only
the first location.
}
\examples{
\dontrun{
data <- export_score(keyword = "amazon")
plot_ts(data)
}

}
