% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_voi_box.r
\name{plot_voi_box}
\alias{plot_voi_box}
\title{Boxplot of VOI time series}
\usage{
plot_voi_box(data_voi, type = "obs")
}
\arguments{
\item{data_voi}{Data exported from \code{export_voi} function.}

\item{type}{Object of class \code{character} indicating the type of time
series-column from data_score that is used for DOI computation, takes
either "obs", "sad", or "trd". Defaults to \code{"obs"}.}
}
\value{
Boxplot of time series as \code{ggplot2} object.
}
\description{
Boxplot of VOI time series
}
\details{

}
\section{Warning}{

\code{plot_voi_box} is limited to 9 unique keywords to avoid an overcrowded
plot. When \code{data_voi} includes more than 9 unique keywords, only
the first 9 keywords are used.
}

\examples{
\dontrun{
data <- export_voi(object = 1)
plot_voi_box(data_voi = data, type = "obs")
plot_voi_box(data_voi = data, type = "sad")
plot_voi_box(data_voi = data, type = "trd")
}

}
