% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_agg.r
\name{run_agg}
\alias{run_agg}
\title{Aggregate keyword-country data and compute DOI}
\usage{
run_agg(control, object, lst_geo = "lst_wdi")
}
\arguments{
\item{control}{Control batch for which the search score is used. Object
of class \code{numeric}.}

\item{object}{Object batch for which the keyword-country data
is aggregated and DOI is computed.  Object of class \code{numeric}.}

\item{lst_geo}{List of locations for which the search score is used.
Object of class \code{character}.}
}
\value{
Message that data was aggregated successfully. Data is uploaded
to data_agg.
}
\description{
Aggregate keyword-country data and compute DOI
}
\details{

}
\examples{
\dontrun{
data_agg(control = 1, object = 1, lst_geo = "lst_wdi")
}

}
\seealso{

}
