% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_control_db.r
\name{run_control}
\alias{run_control}
\title{Download data for control batch}
\usage{
run_control(control, lst_geo = lst_wdi)
}
\arguments{
\item{control}{Control batch for which the data is downloaded. Object
of class \code{numeric}.}

\item{lst_geo}{List of countries or regions for which the data is downloaded.
Refers to lists generated in \code{gtrends_base}.}
}
\value{
Message that data was downloaded successfully. Data is uploaded
to data_control.
}
\description{
Download data for control batch
}
\details{

}
\examples{
\dontrun{
data_con(control = 1, lst_geo = lst_wdi)
}

}
\seealso{

}
