% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_object_db.r
\name{run_object}
\alias{run_object}
\title{Download data for object batch}
\usage{
run_object(object, lst_geo = lst_wdi)
}
\arguments{
\item{object}{Object batch for which the data is downloaded. Object
of class \code{numeric}.}

\item{lst_geo}{List of countries or regions for which the data is downloaded.
Refers to lists generated in \code{gtrends_base}.}
}
\value{
Message that data was downloaded successfully. Data is uploaded
to data_object.
}
\description{
Download data for object batch
}
\details{

}
\examples{
\dontrun{
data_obj(object = 1, lst_geo = lst_wdi)
}

}
\seealso{

}
