% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_remove.r
\name{run_remove}
\alias{run_remove}
\title{Remove data from database tables}
\usage{
run_remove(table, batch_c = NULL, batch_o = NULL)
}
\arguments{
\item{table}{Database table from which the batch should be removed.  Object
of class \code{character}.}

\item{batch_c}{Control batch for which the data is removed Object
of class \code{numeric}.}

\item{batch_o}{Object batch for which the data is removed Object
of class \code{numeric}.}
}
\value{
Message that data was removed successfully. Data is removed
from database.
}
\description{
Remove data from database tables
}
\details{

}
\examples{
\dontrun{
run_remove(table = "batch_terms", batch_c = 1)
run_remove(table = "data_map", batch_c = 1, batch_o = 1)
}

}
\seealso{

}
