% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_score_db.r
\name{run_score}
\alias{run_score}
\title{Compute keyword-country search score}
\usage{
run_score(control, object, lst_geo = lst_wdi)
}
\arguments{
\item{control}{Control batch for which the data is downloaded. Object
of class \code{numeric}.}

\item{object}{Object batch for which the data is downloaded. Object
of class \code{numeric}.}

\item{lst_geo}{List of countries or regions for which the data is downloaded.
Refers to lists generated in \code{gtrends_base}.}
}
\value{
Message that data was computed successfully. Data is uploaded
to data_score.
}
\description{
Compute keyword-country search score
}
\details{

}
\examples{
\dontrun{
data_score(control = 1, object = 1, lst_geo = lst_wdi)
}

}
\seealso{

}
