% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_batch_db.r
\name{add_batch}
\alias{add_batch}
\title{Add keyword batches to data_con and data_obj}
\usage{
add_batch(type, keyword, time = "2010-01-01 2020-07-31")
}
\arguments{
\item{type}{Type batch that should be added. Object of class \code{character}
of value "control" or "object".}

\item{keyword}{Keywords that should be added as batch. Vector of class
\code{character} containing a maximum of five keywords.}

\item{time}{Time frame for which the batch data should be loaded. Object of
class \code{character} that takes the from "YYYY-MM-DD YYYY-MM-DD".}
}
\value{
Message that the batch was created successfully. Batch data is
uploaded to batch_terms and batch_time.
}
\description{
Add keyword batches to data_con and data_obj
}
\details{

}
\examples{
\dontrun{
add_batch(
  type = "control", keyword = c("gmail", "youtube"), time =
    "2016-01-01 2019-12-31"
)
add_batch(
  type = "object", keyword = c("facebook", "google"), time =
    "2016-01-01 2019-12-31"
)
}
}
\seealso{

}
