% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_keywords.r
\name{add_control_keyword}
\alias{add_control_keyword}
\alias{add_object_keyword}
\title{Add batch of control or object keywords}
\usage{
add_control_keyword(keyword, time = "2010-01-01 2019-12-31")

add_object_keyword(keyword, time = "2010-01-01 2019-12-31")
}
\arguments{
\item{keyword}{Keywords that should be added as batch. Vector of class
\code{character} or a \code{list} of \code{character} objects.}

\item{time}{Time frame for which the batch data should be loaded. Object of
class \code{character} that takes the from "YYYY-MM-DD YYYY-MM-DD". Defaults
to "2010-01-01 2019-12-31".}
}
\value{
Message that the batch was created successfully. Batch data is
uploaded to batch_keywords and batch_time.
}
\description{
The function adds a batch of keywords and a time period for downloads to the
database. The batches serve as input for all download and computation
functions.
}
\details{
Since Google Trends allows a maximum of five keywords, batches of control
keywords can consist of up to five keywords. Since one control keyword is
added to batches of object keywords for mapping, object batch length is
limited to four keywords. When a \code{character} vector contains more than
four (five) keywords, the vector is split into four-keyword (five-keyword)
batches. A \code{list} must contain \code{character} objects of length four
(five) or less.
}
\examples{
\dontrun{
add_control_keyword(
  keyword = c("gmail", "maps", "translate", "wikipedia", "youtube"),
  time = "2016-01-01 2019-12-31"
)
add_object_keyword(
  keyword = c("apple", "facebook", "google", "microsoft"),
  time = "2016-01-01 2019-12-31"
)

add_control_keyword(
  keyword = c("gmail", "maps", "news", "translate", "weather", "wikipedia", "youtube"),
  time = "2016-01-01 2019-12-31"
)
add_control_keyword(
  keyword = c("amazon", "apple", "facebook", "google", "microsoft", "netflix", "twitter"),
  time = "2016-01-01 2019-12-31"
)

add_control_keyword(
  keyword = list(
    c("gmail", "maps", "news"),
    c("translate", "weather", "wikipedia", "youtube")
  ),
  time = "2016-01-01 2019-12-31"
)
add_control_keyword(
  keyword = list(
    c("amazon", "apple", "facebook", "google"),
    c("microsoft", "netflix", "twitter")
  ),
  time = "2016-01-01 2019-12-31"
)
}
}
\seealso{
\code{\link{batch_keywords}}, \code{\link{batch_time}}
}
