% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_keywords.r
\name{add_control_keyword}
\alias{add_control_keyword}
\alias{add_object_keyword}
\title{Add batches of control or object keywords}
\usage{
add_control_keyword(keyword, time = "2010-01-01 2019-12-31")

add_object_keyword(keyword, time = "2010-01-01 2019-12-31")
}
\arguments{
\item{keyword}{Keywords that should be added as batch. Vector of class
\code{character} or a \code{list} of \code{character} vectors.}

\item{time}{Time frame for which the batch data should be downloaded. Object
of class \code{character} that takes the from "YYYY-MM-DD YYYY-MM-DD".
Defaults to \emph{"2010-01-01 2019-12-31"}.}
}
\value{
Message that the batch has been created successfully. Batch data is
written to tables \emph{batch_keywords} and \emph{batch_time}.
}
\description{
The function adds one or more batches of keywords with a time period for
downloads to the database. The batches serve as input for all download and
computation functions.
}
\details{
Since Google Trends allows a maximum of five keywords for each query, batches
of control keywords can consist of up to five keywords. Since one control
keyword is added to batches of object keywords for mapping, object batch
length is limited to four keywords. When a \code{character} vector contains
more than four (five) keywords, the vector is split into four-keyword
(five-keyword) batches. A \code{list} must contain \code{character} vectors
of length four (five) or less. Each batch of keywords is combined with a time
period for which data will be downloaded. To change the time period for an
existing batch, all downloads and computations must be rerun.
}
\examples{
\dontrun{
add_control_keyword(
  keyword = c("gmail", "maps", "translate", "wikipedia", "youtube"),
  time = "2016-01-01 2019-12-31"
)
add_object_keyword(
  keyword = c("apple", "facebook", "google", "microsoft"),
  time = "2016-01-01 2019-12-31"
)

add_control_keyword(
  keyword = c("gmail", "maps", "news", "translate", "weather", "wikipedia", "youtube"),
  time = "2016-01-01 2019-12-31"
)
add_control_keyword(
  keyword = c("amazon", "apple", "facebook", "google", "microsoft", "netflix", "twitter"),
  time = "2016-01-01 2019-12-31"
)

add_control_keyword(
  keyword = list(
    c("gmail", "maps", "news"),
    c("translate", "weather", "wikipedia", "youtube")
  ),
  time = "2016-01-01 2019-12-31"
)
add_control_keyword(
  keyword = list(
    c("amazon", "apple", "facebook", "google"),
    c("microsoft", "netflix", "twitter")
  ),
  time = "2016-01-01 2019-12-31"
)
}
}
\seealso{
* \code{\link{batch_keywords}}
* \code{\link{batch_time}}
}
