% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_keywords.r
\name{add_synonym}
\alias{add_synonym}
\alias{add_synonym.character}
\alias{add_synonym.list}
\title{Add keyword synonyms}
\usage{
add_synonym(keyword, synonym)

\method{add_synonym}{character}(keyword, synonym)

\method{add_synonym}{list}(keyword, synonym)
}
\arguments{
\item{keyword}{Keyword of type \code{character} for which the synonmys are
added}

\item{synonym}{Synonmy of type \code{character}}
}
\value{

}
\description{
Add keyword synonyms
}
\details{

}
\examples{
\dontrun{
add_synonym(keyword = "fc bayern", synonym = "bayern munich")
}

}
