% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_synonyms.r
\name{aggregate_synonyms}
\alias{aggregate_synonyms}
\title{Aggregate search scores for synonym terms}
\usage{
aggregate_synonyms(control, vacuum = TRUE)
}
\arguments{
\item{control}{Control batch for which the data is downloaded. Object
of type \code{numeric}. Defaults to 1.}

\item{vacuum}{Indicator whether the function should call \code{vacuum_data()}.
Object of type \code{logical}. Defaults to \code{TRUE}.}
}
\value{
Message that the aggregation of synonyms is complete. During the
aggregation process, messages regarding each step. The function shows the
\code{purrr::map_dfr} and \code{purrr:walk} progress bars.
}
\description{
The function aggregates search scores for object keywords across synonyms
defined through \code{add_synonyms}. Users should run this funciton \strong{after}
running \code{compute_score}. Then, search scores will be added and the synonym
terms removed from the \emph{data_score} table. For efficiency reasons, the
function uses \code{purrr::map} to iterate across \code{batch_o}.
}
\details{
The function proceeds in five steps:
\enumerate{
\item Export data from the database using \code{purrr::map_dfr}
\item Aggreagate score data from synonyms to main term
\item Remove batches that include synonyms from the database using \code{purrr::walk}
\item If \code{vacuum = TRUE}, run \code{vacuum_data()} on the database
\item Write aggregated data to the database using \code{purrr::walk}
}

During these steps, \code{map_dfr} and \code{walk} calls return progress information
through \code{.progress = TRUE}.
}
\section{Note}{

The aggregation of synonyms can be very time-consuming when using a large
dataset! Particulalry, the \code{vacuum_data()} call can require substantial time
to run.
}

\examples{
\dontrun{
compute_score(object = 1:2, control = 1)
aggregate_synonyms(control = 1, vacuum = FALSE)
}

}
\seealso{
\itemize{
\item \code{\link[=add_synonym]{add_synonym()}}
\item \code{\link[=vacuum_data]{vacuum_data()}}
}
}
