% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{batch_keywords}
\alias{batch_keywords}
\title{batch_keywords}
\format{
A tibble with 4 rows and 3 variables:
\describe{
  \item{type}{\code{character} indicating the type of batch, takes "control" for
  control batches and "object" for object batches}
  \item{batch}{\code{integer} number of the batch}
  \item{keyword}{\code{character} keywords included in the respective batch}
}
}
\usage{
batch_keywords
}
\description{
xxx.
}
\keyword{datasets}
