% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{batch_keywords}
\alias{batch_keywords}
\alias{example_keywords}
\title{batch_keywords}
\format{
A tibble with 19 rows and 3 variables:
\describe{
\item{type}{Column of type \code{character} showing the type of each batch,
takes "control" for control batches and "object" for object batches.}
\item{batch}{Column of type \code{integer} showing the number of each
batch.}
\item{keyword}{Column of type \code{character} showing the keywords
included in each batch.}
}
}
\usage{
example_keywords
}
\description{
The table \emph{batch_keywords} contains the keywords for each batch. Each
line contains one \emph{keyword}, the \emph{type} of the batch (i.e. control
or object) and the id of the \emph{batch} to which the keyword is assigned.
Keywords can be added with the function \code{add_keywords}. The function
\code{start_db} exports the table \emph{batch_keywords} as objects
\code{keywords_control} and \code{keywords_object} to \code{.GlobalEnv}.

Example data for the table \emph{batch_keywords} is available as R object
\code{example_keywords}.
}
\seealso{
\itemize{
\item \code{\link[=add_keyword]{add_keyword()}}
}
}
\keyword{datasets}
