% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{batch_keywords}
\alias{batch_keywords}
\title{batch_keywords}
\format{
A tibble with 19 rows and 3 variables:
\describe{
  \item{type}{\code{character} indicating the type of batch, takes "control"
  for control batches and "object" for object batches}
  \item{batch}{\code{integer} number of the batch}
  \item{keyword}{\code{character} keywords included in the respective batch}
}
}
\usage{
batch_keywords
}
\description{
The table \emph{batch_keywords} contains the keywords per batch. Each line
contains one \emph{keyword}, the \emph{type} of the batch (i.e. control or
object) and the id of the \emph{batch} to which the keyword is assigned.
Keywords can be added through \code{add_keywords}. The function
\code{start_db} exports the table \emph{batch_keywords} as objects
\code{keywords_control} and \code{keywords_object} to \code{.GlobalEnv}.
}
\seealso{
\code{\link{add_keyword}}
}
\keyword{datasets}
