% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{batch_time}
\alias{batch_time}
\title{batch_time}
\format{
A tibble with 2 rows and 3 variables:
\describe{
  \item{type}{\code{character} indicating the type of batch, takes "control" for
  control batches and "object" for object batches}
  \item{batch}{\code{integer} number of the batch}
  \item{time}{\code{character} timeframe for the respective batch as
  "YYYY-MM-DD YYYY-MM-DD"}
}
}
\usage{
batch_time
}
\description{
xxx.
}
\keyword{datasets}
