% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{batch_time}
\alias{batch_time}
\title{batch_time}
\format{
A tibble with 5 rows and 3 variables:
\describe{
  \item{type}{\code{character} indicating the type of batch, takes "control"
  for control batches and "object" for object batches}
  \item{batch}{\code{integer} number of the batch}
  \item{time}{\code{character} timeframe for the respective batch as
  "YYYY-MM-DD YYYY-MM-DD"}
}
}
\usage{
batch_time
}
\description{
The table \emph{batch_time} contains the time frame for which data is
downloaded per batch. Each line contains one \emph{time} frame, the
\emph{type} of the batch (i.e. control or object) and the id of the
\emph{batch} to which the time frame is assigned. Time frames take the form
\code{"YYYY-MM-DD YYYY-MM-DD"}. Time frames are added automatically through
\code{add_keywords}. The function \code{start_db} exports the table
\emph{batch_time} as objects \code{time_control} and \code{time_object} to
\code{.GlobalEnv}.
}
\seealso{
\code{\link{add_keywords}}
}
\keyword{datasets}
