% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_doi.r
\name{compute_doi}
\alias{compute_doi}
\alias{compute_doi.numeric}
\alias{compute_doi.list}
\title{Aggregate keyword-country data and compute DOI}
\usage{
compute_doi(object, control = 1, locations = "countries")

\method{compute_doi}{numeric}(object, control = 1, locations = "countries")

\method{compute_doi}{list}(object, control = 1, locations = "countries")
}
\arguments{
\item{object}{Object batch for which the keyword-country data
is aggregated and DOI is computed.  Object of class \code{numeric}.}

\item{control}{Control batch for which the search score is used. Object
of class \code{numeric}.}

\item{locations}{List of locations for which the search score is used.
Object of class \code{character}.}
}
\value{
Message that data was aggregated successfully. Data is uploaded
to data_doi.
}
\description{
Aggregate keyword-country data and compute DOI
}
\details{

}
\examples{
\dontrun{
compute_doi(object = 1, control = 1, locations = "countries")
compute_doi(object = as.list(1:5), control = 1, locations = "countries")
}

}
\seealso{
\code{\link{data_doi}}, \code{\link[ineq]{ineq}}
}
