% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_score.r
\name{compute_score}
\alias{compute_score}
\alias{compute_score.numeric}
\alias{compute_score.list}
\alias{compute_score_global}
\title{Compute keyword-country search score}
\usage{
compute_score(object, control = 1, locations = countries)

\method{compute_score}{numeric}(object, control = 1, locations = countries)

\method{compute_score}{list}(object, control = 1, locations = countries)

compute_score_global(object, control = 1)
}
\arguments{
\item{object}{Object batch for which the data is downloaded. Object
of class \code{numeric} or object of class \code{list} containing single
objects of class \code{numeric}.}

\item{control}{Control batch for which the data is downloaded. Object
of class \code{numeric}. Defaults to 1.}

\item{locations}{List of countries or regions for which the data is
downloaded. Refers to lists generated in \code{start_db}.}
}
\value{
Message that data was computed successfully. Data is uploaded
to data_score.
}
\description{
Compute keyword-country search score
}
\details{

}
\examples{
\dontrun{
compute_score(object = 1, control = 1, locations = countries)
compute_score(object = as.list(1:5), control = 1, locations = countries)
}

}
\seealso{
\code{\link{data_score}}, \code{\link[stats]{stl}},
\code{\link[forecast]{seasadj}}
}
