% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{countries}
\alias{countries}
\title{countries}
\format{
An object of class \code{character} of length 71.
}
\usage{
countries
}
\description{
A character vector that includes ISO2 codes for all countries with a share in
global GDP >= 0.1\\% in 2018. Data on GDP is retrieved from the World Bank's
World Development Indicators database. The data includes:
\itemize{
\item AE
\item AO
\item AR
\item AT
\item AU
\item BD
\item BE
\item BR
\item CA
\item CH
\item CL
\item CN
\item CO
\item CU
\item CZ
\item DE
\item DK
\item DO
\item DZ
\item EC
\item EG
\item ES
\item ET
\item FI
\item FR
\item GB
\item GR
\item HK
\item HU
\item ID
\item IE
\item IL
\item IN
\item IQ
\item IR
\item IT
\item JP
\item KR
\item KW
\item KZ
\item LK
\item MA
\item MX
\item MY
\item NG
\item NL
\item NO
\item NZ
\item OM
\item PE
\item PH
\item PK
\item PL
\item PR
\item PT
\item QA
\item RO
\item RU
\item SA
\item SD
\item SE
\item SG
\item SK
\item TH
\item TR
\item TW
\item UA
\item US
\item UZ
\item VN
\item ZA
}
}
\seealso{
\itemize{
\item \code{\link[WDI:WDI]{WDI::WDI()}}
}
}
\keyword{datasets}
