% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{countries_wdi}
\alias{countries_wdi}
\title{Country codes and names from WDI}
\format{
A data.frame.
}
\usage{
countries_wdi
}
\description{
A data frame of country/location codes and names as provided by the World
Bank World Development Indicators (WDI). This object is useful for mapping
ISO-style codes to human-readable names when constructing custom location
sets.
}
\seealso{
\link{countries}, \code{\link[=add_locations]{add_locations()}}
}
\keyword{datasets}
