% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{countries_wdi}
\alias{countries_wdi}
\title{countries_wdi}
\format{
An object of class \code{data.frame} with 299 rows and 2 columns.
}
\usage{
countries_wdi
}
\description{
A data.frame that includes ISO2 codes and country names for all countries and
locations in the World Bank's World Development Indicators database.
}
\seealso{
\itemize{
\item \code{\link[WDI:WDI_data]{WDI::WDI_data()}}
}
}
\keyword{datasets}
