% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_doi}
\alias{data_doi}
\title{data_doi}
\format{
A tibble with 4,320 rows and 9 variables:
\describe{
  \item{keyword}{\code{character} keyword for which the data was downloaded}
  \item{date}{date as \code{integer} variable, can be transformed into date
  format with \code{lubridate::as_date}}
  \item{type}{\code{character} indicating the type of time series-column from
  \code{data_score} that is used for DOI computation, takes either
  "socre_obs", "score_sad", or "score_trd"}
  \item{gini}{\code{double} DOI computed as inverted Gini coefficient of
  search score from \code{data_score}}
  \item{hhi}{\code{double} DOI computed as inverted Herfindahl-Hirschman
  index of search score from \code{data_score}}
  \item{entropy}{\code{double} DOI computed as inverted Entropy measure for
  search score from \code{data_score}}
  \item{batch_c}{\code{integer} number of control batch}
  \item{batch_o}{\code{integer} number of object batch}
  \item{locations}{\code{character} list of locations for which the search
  score is used}
}
}
\usage{
data_doi
}
\description{
The table *data_doi* contains the degree of internationalization (doi) for
each object batch. Each line contains the doi computed as inverted *gini*
coefficient, as inverted *hhi*, or inverted *entropy* for each *keyword* in
an object *batch_o* for a given *date* and *type* of search score. The column
*batch_c* indicates the control batch that has been used as baseline search
score. Column *locations* indicates which set of locations was used to
compute the distribution of search scores. Degree of internationalization is
computed automatically written to the table through \code{compute_doi}. The
function \code{start_db} exports the table *data_doi* as database connection
\code{tbl_doi} to \code{.GlobalEnv}. Users can access the database table
through \code{dplyr::tbl}.
}
\details{
The sample data included in \code{data_doi} was simulated based on actual
Google Trends data.
}
\seealso{
\code{\link{compute_doi}}, \code{\link[dplyr]{tbl}}
}
\keyword{datasets}
