% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_global}
\alias{data_global}
\title{data_global}
\format{
A tibble with 418 rows and 4 variables:
\describe{
  \item{keyword}{\code{character} keyword for which the data was downloaded}
  \item{date}{date as \code{integer} variable, can be transformed into date
  format with \code{lubridate::as_date}}
  \item{hits}{\code{integer} worldwide search hits for the respective
  keyword-date combination}
  \item{batch}{\code{integer} number of batch}
}
}
\source{
\url{https://trends.google.com/trends/}
}
\usage{
data_global
}
\description{
xxx.
}
\keyword{datasets}
