% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_object}
\alias{data_object}
\title{data_object}
\format{
A tibble with 2090 rows and 5 variables:
\describe{
  \item{location}{\code{character} ISO2 code of country or region for which the
  data was downloaded}
  \item{keyword}{\code{character} keyword for which the data was downloaded}
  \item{date}{date as \code{integer} variable, can be transformed into date
  format with \code{lubridate::as_date}}
  \item{hits}{\code{integer} search hits for the respective location-keyword-date
  combination}
  \item{batch}{\code{integer} number of batch}
}
}
\source{
\url{https://trends.google.com/trends/}
}
\usage{
data_object
}
\description{
xxx.
}
\keyword{datasets}
