% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_score}
\alias{data_score}
\title{data_score}
\format{
A tibble with 480 rows and 8 variables:
\describe{
  \item{location}{\code{character} ISO2 code of country or region for which the
  data was downloaded}
  \item{keyword}{\code{character} keyword for which the data was downloaded}
  \item{date}{date as \code{integer} variable, can be transformed into date
  format with \code{lubridate::as_date}}
  \item{score_obs}{\code{double} search score for the respective location-keyword-date
  combination - no timeseries adjustment}
  \item{score_sad}{\code{double} search score for the respective location-keyword-date
  combination - seasonally adjusted timeseries}
  \item{score_trd}{\code{double} search score for the respective location-keyword-date
  combination - trend-only timeseries}
  \item{batch_c}{\code{integer} number of control batch}
  \item{batch_o}{\code{integer} number of object batch}
}
}
\usage{
data_score
}
\description{
xxx.
}
\keyword{datasets}
