% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.r
\name{disconnect_db}
\alias{disconnect_db}
\title{Disconnect from database}
\usage{
disconnect_db(db = globaltrends_db)
}
\arguments{
\item{db}{Connection to database file that should be closed. Defaults
to \code{globaltrends_db}.}
}
\value{
Message that disconnection was successful.
}
\description{
The function closes the connection to the database file
\emph{db/globaltrends_db.sqlite} in the working directory.
}
\examples{
\dontrun{
disconnect_db()
}
}
\seealso{
* \code{\link{initialize_db}}
* \code{\link{start_db}}
}
