% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_db.r
\name{disconnect_db}
\alias{disconnect_db}
\title{Disconnect from database}
\usage{
disconnect_db(db = doiGT_DB)
}
\value{
Message that disconnection was successful.
}
\description{
Disconnect from database
}
\details{

}
\examples{
\dontrun{
disconnect_db()
}
}
\seealso{

}
