% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.r
\name{.export_data_single}
\alias{.export_data_single}
\alias{.export_data_double}
\title{Run export data from database tables}
\usage{
.export_data_single(
  table,
  in_keyword = NULL,
  in_object = NULL,
  in_control = NULL,
  in_locations = NULL,
  in_type = NULL
)

.export_data_double(
  table,
  in_keyword = NULL,
  in_object = NULL,
  in_control = NULL,
  in_locations = NULL,
  in_type = NULL
)
}
\description{
Run export data from database tables
}
\keyword{internal}
