% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr_test_empty.r
\name{.test_empty}
\alias{.test_empty}
\title{Test wheter table has entries for given criteria}
\usage{
.test_empty(
  table,
  batch_c = NULL,
  batch_o = NULL,
  location = NULL,
  locations = NULL
)
}
\description{
Test wheter table has entries for given criteria
}
\keyword{internal}
