% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_control.r
\name{download_control}
\alias{download_control}
\alias{download_control.numeric}
\alias{download_control.list}
\alias{download_control_global}
\title{Download data for control keywords}
\usage{
download_control(control, locations = countries)

\method{download_control}{numeric}(control, locations = countries)

\method{download_control}{list}(control, locations = countries)

download_control_global(control = 1)
}
\arguments{
\item{control}{Control batch for which the data is downloaded. Object
of class \code{numeric} or object of class \code{list} containing single
objects of class \code{numeric}.}

\item{locations}{List of countries or regions for which the data is downloaded.
Refers to lists generated in \code{start_db}.}
}
\value{
Message that data was downloaded successfully. Data is uploaded
to data_control.
}
\description{
Download data for control keywords
}
\details{

}
\examples{
\dontrun{
download_control(control = 1, locations = countries)
download_control(control = as.list(1:5), locations = countries)
}

}
\seealso{

}
