% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_control.r
\name{download_control}
\alias{download_control}
\alias{download_control.numeric}
\alias{download_control.list}
\title{Download data for control batch}
\usage{
download_control(control, locations = lst_wdi)

\method{download_control}{numeric}(control, locations = lst_wdi)

\method{download_control}{list}(control, locations = lst_wdi)
}
\arguments{
\item{control}{Control batch for which the data is downloaded. Object
of class \code{numeric} or object of class \code{list} containing single
elements of class \code{numeric}.}

\item{locations}{List of countries or regions for which the data is downloaded.
Refers to lists generated in \code{start_db}.}
}
\value{
Message that data was downloaded successfully. Data is uploaded
to data_control.
}
\description{
Download data for control batch
}
\details{

}
\examples{
\dontrun{
data_con(control = 1, locations = lst_wdi)
data_con(control = as.list(1:5), locations = lst_wdi)
}

}
\seealso{

}
