% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_control.r
\name{download_control}
\alias{download_control}
\alias{download_control.numeric}
\alias{download_control.list}
\alias{download_control_global}
\title{Download data for control keywords}
\usage{
download_control(control, locations = countries)

\method{download_control}{numeric}(control, locations = countries)

\method{download_control}{list}(control, locations = countries)

download_control_global(control)
}
\arguments{
\item{control}{Control batch for which the data is downloaded. Object
of class \code{numeric} or object of class \code{list} containing single
objects of class \code{numeric}.}

\item{locations}{List of countries or regions for which the data is
downloaded.' Refers to lists generated in \code{start_db}.}
}
\value{
Message that data was downloaded successfully. Data is uploaded
to data_control.
}
\description{
The function downloads search volumes from Google Trends for a *control*
batch in a set of *locations*. Data is automatically written to table
*data_control*. For \code{download_control_global} the input *location* is
automatically set to *world*.
}
\details{
Downloads through the Google Trends API are made through
\code{gtrendsR::gtrends}. Each control batch can consist of up to five
keywords and is predefined in tables *batch_keywords* and *batch_time*
through \code{add_keywords}. The download for a single keyword batch for a
single location takes about 30 seconds. This includes a randomized waiting
period of 20-30 seconds between downloads. Depending on the frequency of
downloads, Google Trends might block users for some time. In this case,
\code{download_control} waits 60 minutes before it retries the download.
}
\examples{
\dontrun{
download_control(control = 1, locations = countries)
download_control(control = as.list(1:5), locations = countries)
}

}
\seealso{
\code{\link{data_control}}, \code{\link[gtrendsR]{gtrends}}
}
