% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_object.r
\name{download_object}
\alias{download_object}
\alias{download_object.numeric}
\alias{download_object.list}
\title{Download data for object batch}
\usage{
download_object(object, locations = lst_wdi)

\method{download_object}{numeric}(object, locations = lst_wdi)

\method{download_object}{list}(object, locations = lst_wdi)
}
\arguments{
\item{object}{Object batch for which the data is downloaded. Object
of class \code{numeric} or object of class \code{list} containing single
elements of class \code{numeric}.}

\item{locations}{List of countries or regions for which the data is downloaded.
Refers to lists generated in \code{start_db}.}
}
\value{
Message that data was downloaded successfully. Data is uploaded
to data_object.
}
\description{
Download data for object batch
}
\details{

}
\examples{
\dontrun{
data_obj(object = 1, locations = lst_wdi)
data_obj(object = as.list(1:5), locations = lst_wdi)
}

}
\seealso{

}
