% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_object.r
\name{download_object}
\alias{download_object}
\alias{download_object.numeric}
\alias{download_object.list}
\title{Download data for object batch}
\usage{
download_object(object, control = 1, locations = countries)

\method{download_object}{numeric}(object, control = 1, locations = countries)

\method{download_object}{list}(object, control = 1, locations = countries)
}
\arguments{
\item{object}{Object batch for which the data is downloaded. Object
of class \code{numeric} or object of class \code{list} containing single
elements of class \code{numeric}.}

\item{control}{Control batch that is used for mapping. Object of class
\code{numeric}. Defaults to \code{1}.}

\item{locations}{List of countries or regions for which the data is downloaded.
Refers to lists generated in \code{start_db}.}
}
\value{
Message that data was downloaded successfully. Data is uploaded
to data_objectect.
}
\description{
Download data for object batch
}
\details{

}
\examples{
\dontrun{
download_object(object = 1, locations = countries)
download_object(object = as.list(1:5), locations = countries)
}

}
\seealso{

}
