% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_object.r
\name{download_object}
\alias{download_object}
\alias{download_object.numeric}
\alias{download_object.list}
\alias{download_object_global}
\title{Download data for object batch}
\usage{
download_object(object, control = 1, locations = countries)

\method{download_object}{numeric}(object, control = 1, locations = countries)

\method{download_object}{list}(object, control = 1, locations = countries)

download_object_global(object, control = 1)
}
\arguments{
\item{object}{Object batch for which the data is downloaded. Object
of class \code{numeric} or object of class \code{list} containing single
object of class \code{numeric}.}

\item{control}{Control batch that is used for mapping. Object of class
\code{numeric}. Defaults to \code{1}.}

\item{locations}{List of countries or regions for which the data is
downloaded. Refers to lists generated in \code{start_db}.}
}
\value{
Message that data was downloaded successfully. Data is uploaded
to data_object.
}
\description{
The function downloads search volumes from Google Trends for an object batch
(\emph{batch_o}) and one keyword from a control batch (\emph{batch_c}) in a
set of \emph{locations}. Data is automatically written to table
\emph{data_object}. For \code{download_object_global} the input
\emph{location} is automatically set to \emph{world}.
}
\details{
Downloads through the Google Trends API are made through
\code{gtrendsR::gtrends}. Each object batch can consist of up to four
keywords and is predefined in tables *batch_keywords* and *batch_time*
through \code{add_keywords}. In addition, one control keyword is added to
each object batch. The control keyword then allows a mapping between search
volumes for control keywords stored in *data_control* and search volumes for
object keywords. The download for a single keyword batch for a single
location takes about 30 seconds. This includes a randomized waiting period of
20-30 seconds between downloads. Depending on the frequency of downloads,
Google Trends might block users for some time. In this case,
\code{download_object} waits 60 minutes before it retries the download.
}
\examples{
\dontrun{
download_object(object = 1, locations = countries)
download_object(object = as.list(1:5), locations = countries)
}

}
\seealso{
\code{\link{data_object}}, \code{\link[gtrendsR]{gtrends}}
}
