% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.r
\name{export_control}
\alias{export_control}
\alias{export_object}
\alias{export_global}
\alias{export_mapping}
\alias{export_score}
\alias{export_doi}
\title{Export data from data_control table}
\usage{
export_control(control = NULL)

export_object(keyword = NULL, object = NULL)

export_global(keyword = NULL, object = NULL)

export_mapping(keyword = NULL, object = NULL, control = NULL, locations = NULL)

export_score(keyword = NULL, object = NULL, control = NULL, locations = NULL)

export_doi(
  keyword = NULL,
  object = NULL,
  control = NULL,
  locations = NULL,
  type = NULL
)
}
\arguments{
\item{control}{Control batch number for which data should be exported. Element
of class \code{numeric}. When element length exceeds 1, only the first element
is used for exporting.}

\item{keyword}{Object keywords for which data should be exported. Element of
class \code{character}. When element length exceeds 1, only the first element
is used for exporting.}

\item{object}{Object batch number for which data should be exported. Element
of class \code{numeric}. When element length exceeds 1, only the first element
is used for exporting.}

\item{locations}{Set of locations for which data should be exported. Element
of class \code{character}. When element length exceeds 1, only the first
element is used for exporting.}

\item{type}{Type of timeseries for which data should be exported. Element
of class \code{character}. When element length exceeds 1, only the first
element is used for exporting. Relevant only for "export_doi". Takes one
of the following values: "score_obs", "score_sad", "score_trd".}
}
\value{
The functions export and filter the respective database tables and return
objects of class \code{"tbl_df", "tbl", "data.frame"}. In addition, the
each function adds a specific class that can be used for further methods.
\itemize{
\item \code{export_control} exports table of class \code{doiGT_control}
with columns location, keyword, date, hits, control
\item \code{export_object} exports table of class \code{doiGT_object}
with columns location, keyword, date, hits, object
\item \code{export_global} exports table of class \code{doiGT_global}
with columns keyword, date, hits, control
\item \code{export_mapping} exports table of class \code{doiGT_mapping}
with columns location, keyword, date, hits, control, object
\item \code{export_score} exports table of class \code{doiGT_score}
with columns location, keyword, date, score_obs, score_sad, score_trd,
control, object
\item \code{export_doi} exports table of class \code{doiGT_doi}
with columns keyword, date, type, gini, hhi, entropy, control, object,
locations
}
}
\description{
Export data from data_control table

Export data from data_object table

Export data from data_global table

Export data from data_mapping table

Export data from data_score table

Export data from data_doi table
}
\details{

}
\examples{
\dontrun{
export_control(control = 2)

export_object(keyword = "manchester united")

export_score(object = 3, control = 1, locations = "countries") \%>\%
  readr::write_csv("data_score.csv")

export_doi(keyword = "manchester united", control = 2, type = "score_sad", locations = "us_states") \%>\%
  write_xl::write_xlsx("data_doi.xlsx")
}

}
\seealso{

}
