% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.r
\name{export_control}
\alias{export_control}
\alias{export_control_global}
\alias{export_object}
\alias{export_object_global}
\alias{export_score}
\alias{export_voi}
\alias{export_doi}
\title{Export data from database table}
\usage{
export_control(control = NULL, locations = NULL)

export_control_global(control = NULL)

export_object(keyword = NULL, object = NULL, control = NULL, locations = NULL)

export_object_global(keyword = NULL, object = NULL, control = NULL)

export_score(keyword = NULL, object = NULL, control = NULL, locations = NULL)

export_voi(keyword = NULL, object = NULL, control = NULL)

export_doi(
  keyword = NULL,
  object = NULL,
  control = NULL,
  locations = NULL,
  type = NULL
)
}
\arguments{
\item{control}{Control batch number for which data should be exported.}

\item{locations}{List of locations for which the search score is used.
For \code{export_control}, \code{export_object}, or \code{export_score}
refers to lists generated in \code{start_db}. For \code{export_doi}
object of class \code{character}.}

\item{keyword}{Object keywords for which data should be exported. Object of
class \code{character}.}

\item{object}{Object batch number for which data should be exported.}

\item{type}{Type of time series for which data should be exported. Element
of class \code{character}. Relevant only for \code{export_global} and
\code{export_doi}. Takes one of the following values: \emph{obs} - observed
search scores, \emph{sad} - seasonally adjusted search scores, \emph{trd} -
trend only search scores.}
}
\value{
The functions export and filter the respective database tables and return
objects of class \code{"tbl_df", "tbl", "data.frame"}.
\itemize{
  \item \code{export_control} exports data from table \emph{data_control} with
columns location, keyword, date, hits, control.
  \item \code{export_object} exports data from table \emph{data_object} with
columns location, keyword, date, hits, object.
  \item \code{export_score} exports data from table \emph{data_score} with
columns location, keyword, date, score_obs, score_sad, score_trd, control,
object. Object of class \code{c("exp_score", "data.frame")}.
  \item \code{export_voi} exports data from table \emph{data_score} with
columns keyword, date, hits, control, filters for
\code{location == "world"}. Object of class
\code{c("exp_voi", "data.frame")}.
  \item \code{export_doi} exports data from table \emph{data_doi} with columns
keyword, date, type, gini, hhi, entropy, control, object, locations. Object
of class \code{c("exp_doi", "data.frame")}.
}
}
\description{
The function allows to export data from database tables. In combination with
various \emph{write} functions in R, the functions allow exports from the
database to local files.
}
\details{
Exports can be filtered by \emph{keyword}, \emph{object}, \emph{control},
\emph{locations}, or \emph{type}. Not all filters are applicable for all
functions. When filter \emph{keyword} and \emph{object} are used together,
\emph{keyword} overrules \emph{object}. Currently the functions do not
include list inputs - users are advised to \code{purrr::map_dfr} or
\code{dplyr::filter} instead.
}
\examples{
\dontrun{
export_control(control = 2)

export_object(
  keyword = "manchester united",
  locations = countries
)

export_score(
  object = 3,
  control = 1,
  locations = us_states
) \%>\%
  readr::write_csv("data_score.csv")

export_doi(
  keyword = "manchester united",
  control = 2,
  type = "sad",
  locations = "us_states"
) \%>\%
  write_xl::write_xlsx("data_doi.xlsx")

# interaction with purrr::map_dfr
purrr::map_dfr(
  c("coca cola", "microsoft"),
  export_doi,
  control = 1,
  type = "obs"
)

# interaction with dplyr::filter
export_voi(
  object = 1,
  control = 1,
  type = "obs"
) \%>\%
  dplyr::filter(lubridate::year(date) == 2019)
}

}
\seealso{
* \code{\link{data_control}}
* \code{\link{data_object}}
* \code{\link{data_score}}
* \code{\link{data_doi}}
* \code{\link[purrr]{map}}
* \code{\link[dplyr]{filter}}
}
