% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.r
\name{export_control}
\alias{export_control}
\alias{export_object}
\alias{export_global}
\alias{export_score}
\alias{export_doi}
\title{Export data from database table}
\usage{
export_control(control = NULL)

export_object(keyword = NULL, object = NULL)

export_global(keyword = NULL, object = NULL, type = NULL)

export_score(keyword = NULL, object = NULL, control = NULL)

export_doi(
  keyword = NULL,
  object = NULL,
  control = NULL,
  locations = NULL,
  type = NULL
)
}
\arguments{
\item{control}{Control batch number for which data should be exported. Element
of class \code{numeric}. When element length exceeds 1, only the first element
is used for exporting.}

\item{keyword}{Object keywords for which data should be exported. Element of
class \code{character}. When element length exceeds 1, only the first element
is used for exporting.}

\item{object}{Object batch number for which data should be exported. Element
of class \code{numeric}. When element length exceeds 1, only the first element
is used for exporting.}

\item{type}{Type of time series for which data should be exported. Element
of class \code{character}. When element length exceeds 1, only the first
element is used for exporting. Relevant only for "export_global" and
"export_doi". Takes one of the following values: "obs", "sad", "trd".}

\item{locations}{Set of locations for which data should be exported. Element
of class \code{character}. When element length exceeds 1, only the first
element is used for exporting.}
}
\value{
The functions export and filter the respective database tables and return
objects of class \code{"tbl_df", "tbl", "data.frame"}.
\itemize{
\item \code{export_control} exports table with columns location, keyword,
date, hits, control
\item \code{export_object} exports table  with columns location, keyword,
date, hits, object
\item \code{export_global} exports table  with columns keyword, date, hits,
control
\item \code{export_score} exports table  with columns location, keyword,
date, score_obs, score_sad, score_trd, control, object
\item \code{export_doi} exports table with columns keyword, date, type, gini,
hhi, entropy, control, object, locations
}
}
\description{
Export data from database table
}
\details{

}
\examples{
\dontrun{
export_control(control = 2)

export_object(keyword = "manchester united")

export_score(object = 3, control = 1) \%>\%
  readr::write_csv("data_score.csv")

export_doi(
  keyword = "manchester united", control = 2, type = "sad",
  locations = "us_states"
) \%>\%
  write_xl::write_xlsx("data_doi.xlsx")
}

}
\seealso{

}
