% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.r
\name{initialize_db}
\alias{initialize_db}
\title{Initialize database}
\usage{
initialize_db()
}
\value{
Database is created.
}
\description{
Initialize database
}
\examples{
\dontrun{
initialize_db()
}

}
\seealso{

}
