% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_python.r
\name{initialize_python}
\alias{initialize_python}
\title{Initialize gtrends Python module}
\usage{
initialize_python(api_key, conda_env = NULL, python_env = NULL)
}
\arguments{
\item{api_key}{API key obtained from Google. Object of type \code{character}.}

\item{conda_env}{Storage location of a Conda environment. Object of type
\code{character}.}

\item{python_env}{Storage location of a Python environment. Object of type
\code{character}.}
}
\value{
The functions sets the environment variable \code{gt.env} to \code{TRUE}. This channels
downloads to the API. Otherwise, \code{globaltrends} uses \code{gtrendsR::gtrends} for
downloads.
}
\description{
The function initializes the Python session required to download data from
Google Trends using the Google Trends research API. The user can define an
API key and either a Conda or Python environment.
}
\details{
To use this API acces, you must:
\enumerate{
\item Apply for API access \href{https://support.google.com/trends/contact/trends_api}{online}.
and generate an API key in your Google developer console.
\item Create a Python or Conda environment where you install the
\code{google-api-python-client} package with pip.
}
}
\examples{
\dontrun{
initialize_python(
  api_key = "XXX",
  conda_env = "/xxx/venv"
)
}

}
