% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.r
\name{plot_bar}
\alias{plot_bar}
\alias{plot_bar.exp_score}
\title{Create barplot for cross-sectional globaltrends data}
\usage{
plot_bar(data, ...)

\method{plot_bar}{exp_score}(data, ...)
}
\arguments{
\item{data}{Data exported from \code{export_...} functions.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Barplot of cross-sectional data as \code{ggplot2} object.
}
\description{
The function creates barplots for cross-sectional search score data. It uses
the output of \code{export_score} to prepare a bar plot of search scores for
the top 10 countries. When the output includes more than one keyword,
only the first keyword is used.
}
\examples{
\dontrun{
data <- export_score(keyword = "amazon")
plot_bar(data)

data <- export_score(keyword = "amazon")
plot_bar(data)
}

}
