% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.r
\name{plot_box}
\alias{plot_box}
\alias{plot_box.exp_score}
\alias{plot_box.exp_voi}
\alias{plot_box.exp_doi}
\title{Create boxplot for time series of globaltrends data}
\usage{
plot_box(data, ...)

\method{plot_box}{exp_score}(data, ...)

\method{plot_box}{exp_voi}(data, ...)

\method{plot_box}{exp_doi}(
  data,
  measure = c("gini", "hhi", "entropy"),
  locations = "countries",
  ...
)
}
\arguments{
\item{data}{Data exported from \code{export_...} functions.}

\item{...}{Further arguments passed to or from other methods.}

\item{measure}{Object of type \code{character} indicating the DOI measure,
takes either \emph{gini}, \emph{hhi}, or \emph{entropy}. Defaults to \emph{"gini"}.}

\item{locations}{Object of type \code{character} indicating for which
set of locations should be filtered. Defaults to \emph{"countries"}.}
}
\value{
Boxplot of time series as \code{ggplot2} object.
}
\description{
The function creates boxplots for time series globaltrends data. It uses
the output of \code{export_...} to prepare boxplots for up to nine keywords.
}
\details{
For output of \code{export_score}, only data for a single location is shown.
When date for more than one location is provided, the function selects only
the first location.
}
\examples{
\dontrun{
data <- export_score(keyword = "amazon")
plot_box(data)

data <- export_voi(keyword = "amazon")
plot_box(data)

data <- export_doi(keyword = "amazon")
plot_box(data, ci = 0.9)
}

}
