% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_doi_ts.r
\name{plot_doi_ts}
\alias{plot_doi_ts}
\title{Line plot of DOI time series}
\usage{
plot_doi_ts(
  data_doi,
  type = "obs",
  measure = "gini",
  locations = "countries",
  smooth = TRUE
)
}
\arguments{
\item{data_doi}{Data exported from \code{export_doi} function.}

\item{type}{Object of class \code{character} indicating the type of time
series-column from data_score that is used for DOI computation, takes
either \emph{obs}, \emph{sad}, or \emph{trd}. Defaults to \emph{"obs"}.}

\item{measure}{Object of class \code{character} indicating the measure
used for DOI computation, takes either \emph{gini}, \emph{hhi}, or
\emph{entropy}. Defaults to \emph{"gini"}.}

\item{locations}{Object of class \code{character} indicating for which
set of locations should be filtered. Defaults to \emph{"countries"}.}

\item{smooth}{Object of class \code{logical} indicating whether the
\code{geom_smooth} function of \code{ggplot2} should be used. Defaults to
\code{TRUE}.}
}
\value{
Line plot of DOI time series as \code{ggplot2} object.
}
\description{
The function uses the output of \code{export_doi} to prepare a time series
plot of degree of internationalization values. When the output includes more
than nine keywords, only the first nine keywords are used.
}
\examples{
\dontrun{
data <- export_doi(
  object = 1,
  locations = "countries"
)
plot_doi_ts(
  data_doi = data,
  type = "obs",
  measure = "gini",
  smooth = TRUE
)
plot_doi_ts(
  data_doi = data,
  type = "sad",
  measure = "hhi",
  smooth = FALSE
)
plot_doi_ts(
  data_doi = data,
  type = "trd",
  measure = "entropy",
  smooth = TRUE
)
}

}
\seealso{
* \code{\link{export_doi}}
* \code{\link[ggplot2]{ggplot}}
}
