% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score.r
\name{plot_score}
\alias{plot_score}
\title{Bar plot of data_score distribution}
\usage{
plot_score(data_score, type = "obs")
}
\arguments{
\item{data_score}{Data exported from \code{export_score} function.}

\item{type}{Object of class \code{character} indicating the type of time
series-column from data_score that is used, takes either "obs", "sad", or
"trd". Defaults to "obs".}
}
\value{
Bar plot of score distribution across locations as \code{ggplot2}
object.
}
\description{
The function uses the output of \code{export_score} to prepare a bar plot of
search scores for the top 10 countries. When the output includes more than
one keyword, only the first keyword is used.
}
\examples{
\dontrun{
data <- export_score(keyword = "manchester united", locations = "countries")
plot_score(data_score = data, type = "obs")
plot_score(data_score = data, type = "sad")
data <- dplyr::filter(data, location \%in\% countries)
plot_score(data_score = data, type = "trd")
}

}
\seealso{
\code{\link{export_score}}
}
