% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.r
\name{plot_ts}
\alias{plot_ts}
\title{Line plot of data_doi timeseries}
\usage{
plot_ts(
  data_doi,
  type = NULL,
  measure = "gini",
  locations = NULL,
  grid = TRUE,
  smooth = TRUE
)
}
\arguments{
\item{data_doi}{Data exported from \code{export_doi} function.}

\item{type}{Object of class \code{character} indicating the type of time
series-column from data_score that is used for DOI computation, takes
either "obs", "sad", or "trd". By default takes \code{NULL}
and assumes that only one single type of timeseries is included in the
data.}

\item{measure}{Object of class \code{character} indicating the measure
used for DOI computation, takes either "gini", "hhi", or "entropy".
Defaults to "gini".}

\item{locations}{Object of class \code{character} indicating for which
set of locations should be filtered. By default takes \code{NULL} and
assumes that only one single set of locations is included in the data.}

\item{grid}{Object of class \code{logical} indicating whether the
\code{facet_wrap} function of \code{ggplot2} should be used. Otherwise,
keywords are separated by color.}

\item{smooth}{Object of class \code{logical} indicating whether the
\code{geom_smooth} function of \code{ggplot2} should be used. Only
applicale when \code{grid == TRUE}.}
}
\value{
Line plot of timeseries as \code{ggplot2} object
}
\description{
Line plot of data_doi timeseries
}
\details{

}
\section{Warning}{

\code{plot_ts} is limited to 9 unique keywords for \code{grid == TRUE}
and 5 unique keywords otherwise to avoid an overcrowded plot. When
\code{data_doi} includes more than 9 or 5 unique keywords, only
the first 9 or 5 keywords are used.
}

\examples{
\dontrun{
data <- export_doi(object = 1, locations = "countries")
plot_ts(data_doi = data, type = "obs", measure = "gini", grid = TRUE, smooth = TRUE)
plot_ts(data_doi = data, type = "sad", measure = "hhi", grid = TRUE, smooth = FALSE)
plot_ts(data_doi = data, type = "trd", measure = "entropy", grid = FALSE, smooth = TRUE)
}

}
