% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.r
\name{plot_ts}
\alias{plot_ts}
\alias{plot_ts.exp_score}
\alias{plot_ts.abnorm_score}
\alias{plot_ts.exp_voi}
\alias{plot_ts.abnorm_voi}
\alias{plot_ts.exp_doi}
\alias{plot_ts.abnorm_doi}
\title{Plot time series of globaltrends data}
\usage{
plot_ts(data, ...)

\method{plot_ts}{exp_score}(data, type = "obs", smooth = TRUE)

\method{plot_ts}{abnorm_score}(data, ci = 0.95)

\method{plot_ts}{exp_voi}(data, type = "obs", smooth = TRUE)

\method{plot_ts}{abnorm_voi}(data, ci = 0.95)

\method{plot_ts}{exp_doi}(
  data,
  type = "obs",
  measure = "gini",
  locations = "countries",
  smooth = TRUE
)

\method{plot_ts}{abnorm_doi}(data, type = "obs", locations = "countries", ci = 0.95)
}
\arguments{
\item{data}{Data exported from \code{export_...} or \code{compute_abnorm}
functions.}

\item{type}{Object of type \code{character} indicating the type of time
series-column from data_score, takes either \emph{obs}, \emph{sad}, or
\emph{trd}. Defaults to \emph{"obs"}.}

\item{smooth}{Object of type \code{logical} indicating whether the
\code{geom_smooth} function of \code{ggplot2} should be used. Defaults to
\code{TRUE}.}

\item{ci}{Confidence interval within which changes are assumed to be normal.
Object of type \code{double, 0 < ci < 1}. Defaults to \emph{0.95}.}

\item{measure}{Object of type \code{character} indicating the DOI measure,
takes either \emph{gini}, \emph{hhi}, or \emph{entropy}. Defaults to
\emph{"gini"}.}

\item{locations}{Object of type \code{character} indicating for which
set of locations should be filtered. Defaults to \emph{"countries"}.}
}
\value{
Line plot of time series as \code{ggplot2} object. For objects of class
\code{abnorm_...} changes with percentile ranks outside the provided
confidence interval are indicated by red dots.
}
\description{

}
