% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_voi_box.r
\name{plot_voi_box}
\alias{plot_voi_box}
\title{Boxplot of VOI time series}
\usage{
plot_voi_box(data_voi, type = "obs")
}
\arguments{
\item{data_voi}{Data exported from \code{export_voi} function.}

\item{type}{Object of class \code{character} indicating the type of time
series-column from data_score that is used for DOI computation, takes
either \emph{obs}, \emph{sad}, or \emph{trd}. Defaults to \emph{"obs"}.}
}
\value{
Boxplot of VOI distribution as \code{ggplot2} object.
}
\description{
The function uses the output of \code{export_voi} to prepare a box plot of
the distribution of volume of internationalization values. When the output
includes more than nine keywords, only the first nine keywords are used.
}
\examples{
\dontrun{
data <- export_voi(object = 1)
plot_voi_box(
  data_voi = data,
  type = "obs"
)
plot_voi_box(
  data_voi = data,
  type = "sad"
)
plot_voi_box(
  data_voi = data,
  type = "trd"
)
}

}
\seealso{
\code{\link{export_voi}}
}
