% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_data.r
\name{remove_data}
\alias{remove_data}
\title{Remove data from database tables}
\usage{
remove_data(table, control = NULL, object = NULL)
}
\arguments{
\item{table}{Database table from which the batch should be removed.  Object
of class \code{character}.}

\item{control}{Control batch for which the data is removed Object
of class \code{numeric}.}

\item{object}{Object batch for which the data is removed Object
of class \code{numeric}.}
}
\value{
Message that data was removed successfully. Data is removed
from database.
}
\description{
Remove data from database tables
}
\details{

}
\examples{
\dontrun{
remove_data(table = "batch_keywords", control = 1)
remove_data(table = "data_mapping", control = 1, object = 1)
}

}
\seealso{

}
