% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wrld_db.r
\name{run_wrld}
\alias{run_wrld}
\title{Download worlwide score data for object batch}
\usage{
run_wrld(object)
}
\arguments{
\item{object}{Object batch for which the data is downloaded. Object
of class \code{numeric}.}
}
\value{
Message that data was downloaded successfully. Data is uploaded
to data_wrld.
}
\description{
Download worlwide score data for object batch
}
\details{

}
\examples{
\dontrun{
data_wrld(object = 1)
}

}
\seealso{

}
