% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{us_states}
\alias{us_states}
\title{Default location set: US states}
\format{
A character vector.
}
\usage{
us_states
}
\description{
Character vector of US state-level location codes used by the package.

The vector contains ISO 3166-2 codes of the form \code{"US-XX"} for the 50 US
states and \code{"US-DC"} for the District of Columbia.
}
\seealso{
\code{\link[=add_locations]{add_locations()}}, \code{\link[=start_db]{start_db()}}
}
\keyword{datasets}
