% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{us_states}
\alias{us_states}
\title{us_states}
\format{
An object of class \code{character} of length 51.
}
\usage{
us_states
}
\description{
A character vector that includes ISO2 codes for all US federal states and
Washington DC. The data includes:
\itemize{
\item US-AL
\item US-AK
\item US-AZ
\item US-AR
\item US-CA
\item US-CO
\item US-CT
\item US-DE
\item US-FL
\item US-GA
\item US-HI
\item US-ID
\item US-IL
\item US-IN
\item US-IA
\item US-KS
\item US-KY
\item US-LA
\item US-ME
\item US-MD
\item US-MA
\item US-MI
\item US-MN
\item US-MS
\item US-MO
\item US-MT
\item US-NE
\item US-NV
\item US-NH
\item US-NJ
\item US-NM
\item US-NY
\item US-NC
\item US-ND
\item US-OH
\item US-OK
\item US-OR
\item US-PA
\item US-RI
\item US-SC
\item US-SD
\item US-TN
\item US-TX
\item US-UT
\item US-VT
\item US-VA
\item US-WA
\item US-WV
\item US-WI
\item US-WY
\item US-DC
}
}
\keyword{datasets}
